embedding-service-python/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sentence-transformers>=2.2.2
torch>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0
kafka-python>=2.0.2
redis>=5.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
